Refactor date picker component to ember cli date picker addon.
