+++ blueprints/ember-cli-date-picker/index.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = {
+  name: 'ember-date-picker',
+
+  normalizeEntityName: function() { },
+
+  afterInstall: function() {
+    return this.addBowerPackagesToProject([
+      {name: 'ember-spin-box', target: '~0.0.5'}
+    ]);
+  }
+};
