+++ bower.json
@@ -1,20 +1,17 @@
 {
   "name": "ember-date-picker",
-  "version": "0.0.4",
-  "description": "A lightweight, mobile-optimized, date picker component for ember.js applications.",
-  "author": "Bill Dami <billdami@gmail.com>",
-  "license": "MIT",
-  "ignore": [
-    "**/.*"
-  ],
-  "devDependencies": {
-    "jquery": "~2.0.3",
-    "handlebars": "1.3.0",
-    "ember": "~1.5.0",
-    "qunit": "~1.14.0"
-  },
   "dependencies": {
-    "ember-spin-box": "~0.0.5",
-    "momentjs": "~2.6.0"
+    "handlebars": "~1.3.0",
+    "jquery": "^1.11.1",
+    "ember": "1.8.1",
+    "ember-data": "1.0.0-beta.11",
+    "ember-resolver": "~0.1.10",
+    "loader.js": "stefanpenner/loader.js#1.0.1",
+    "ember-cli-shims": "stefanpenner/ember-cli-shims#0.0.3",
+    "ember-cli-test-loader": "rwjblue/ember-cli-test-loader#0.0.4",
+    "ember-load-initializers": "stefanpenner/ember-load-initializers#0.0.2",
+    "ember-qunit": "0.1.8",
+    "ember-qunit-notifications": "0.0.4",
+    "qunit": "~1.15.0"
   }
 }
