--- example/example.css
@@ -1,6 +1,0 @@
-body {
-	padding: 40px;
-	font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-	font-size: 14px;
-	line-height: 1.42857143;
-}
