--- lib/templates-bottom.js
@@ -1,1 +1,0 @@
-});
