+++ package.json
@@ -1,19 +1,39 @@
 {
   "name": "ember-date-picker",
-  "version": "0.0.4",
-  "description": "A lightweight, mobile-optimized, date picker component for ember.js applications.",
-  "author": "Bill Dami <billdami@gmail.com>",
+  "version": "0.0.1",
+  "directories": {
+    "doc": "doc",
+    "test": "tests"
+  },
+  "scripts": {
+    "start": "ember server",
+    "build": "ember build",
+    "test": "ember test"
+  },
+  "repository": "",
+  "engines": {
+    "node": ">= 0.10.0"
+  },
+  "author": "",
   "license": "MIT",
   "devDependencies": {
-    "broccoli": "~0.11.0",
-    "broccoli-merge-trees": "~0.1.3",
-    "broccoli-uglify-js": "~0.1.3",
-    "broccoli-concat": "0.0.5",
-    "broccoli-less-single": "~0.1.4",
-    "broccoli-template-builder": "0.0.1",
-    "broccoli-env": "0.0.1",
-    "broccoli-static-compiler": "~0.1.4",
-    "handlebars": "~1.3.0",
-    "ember-template-compiler": "~1.6.0-beta.3"
+    "broccoli-asset-rev": "^1.0.0",
+    "broccoli-ember-hbs-template-compiler": "^1.6.1",
+    "ember-cli": "0.1.3",
+    "ember-cli-esnext": "0.1.1",
+    "ember-cli-ic-ajax": "0.1.1",
+    "ember-cli-inject-live-reload": "^1.3.0",
+    "ember-cli-qunit": "0.1.2",
+    "ember-data": "1.0.0-beta.11",
+    "ember-export-application-global": "^1.0.0",
+    "express": "^4.8.5",
+    "glob": "^4.0.5"
+  },
+  "description": "The default blueprint for ember-cli addons.",
+  "keywords": [
+    "ember-addon"
+  ],
+  "ember-addon": {
+    "configPath": "tests/dummy/config"
   }
 }
