--- tests/app.js
@@ -1,4 +1,0 @@
-App = Ember.Application.create();
-App.setupForTesting();
-App.injectTestHelpers();
-App.rootElement = '#ember-testing';
