+++ tests/dummy/app/styles/app.css
@@ -1,0 +1,3 @@
+html, body {
+  margin: 20px;
+}
