+++ tests/dummy/public/robots.txt
@@ -1,0 +1,2 @@
+# http://www.robotstxt.org
+User-agent: *
