+++ tests/index.html
@@ -1,40 +1,43 @@
 <!DOCTYPE html>
-<html lang="en">
-    <head>
-        <title>ember-date-picker</title>
-        <meta charset="UTF-8">
-        <meta name="author" content="Bill Dami" />
-        <link rel="stylesheet" href="../bower_components/ember-spin-box/dist/ember-spin-box.css" media="all">
-        <link rel="stylesheet" href="../dist/ember-date-picker.css" media="all">
-        <link rel="stylesheet" href="../bower_components/qunit/qunit/qunit.css" media="all">
-        <style type="text/css">
-            #ember-testing {
-                height: 480px;
-                margin-top: 20px;
-                padding: 10px;
-                overflow: auto;
-                background: #fff;
-                border: 1px solid #ccc;
-            }
-        </style>
-    </head>
-    <body>
-        <div id="qunit"></div>
-        <div id="ember-testing"></div>
-    
-        <script type="text/x-handlebars">
-            {{date-picker-input controls="my-datepicker" value="May 15, 2014" dateFormat="MMMM D, YYYY"}}
-            {{date-picker-controls id="my-datepicker"}}
-        </script>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title>Dummy Tests</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
 
-        <script src="../bower_components/jquery/jquery.js"></script>
-        <script src="../bower_components/handlebars/handlebars.js"></script>
-        <script src="../bower_components/ember/ember.js"></script>
-        <script src="../bower_components/momentjs/min/moment.min.js"></script>
-        <script src="../bower_components/ember-spin-box/dist/ember-spin-box.js"></script>
-        <script src="../bower_components/qunit/qunit/qunit.js"></script>
-        <script src="../dist/ember-date-picker.js"></script>
-        <script src="./app.js"></script>
-        <script src="./tests.js"></script>
-    </body>
+    {{content-for 'head'}}
+    {{content-for 'test-head'}}
+
+    <link rel="stylesheet" href="assets/vendor.css">
+    <link rel="stylesheet" href="assets/dummy.css">
+    <link rel="stylesheet" href="assets/test-support.css">
+    <style>
+      #ember-testing-container {
+        position: absolute;
+        background: white;
+        bottom: 0;
+        right: 0;
+        width: 640px;
+        height: 384px;
+        overflow: auto;
+        z-index: 9999;
+        border: 1px solid #ccc;
+      }
+      #ember-testing {
+        zoom: 50%;
+      }
+    </style>
+  </head>
+  <body>
+
+    {{content-for 'body'}}
+    {{content-for 'test-body'}}
+    <script src="assets/vendor.js"></script>
+    <script src="assets/test-support.js"></script>
+    <script src="assets/dummy.js"></script>
+    <script src="testem.js"></script>
+    <script src="assets/test-loader.js"></script>
+  </body>
 </html>
